Better default features and project layout
