+++ src/util.rs
@@ -1,7 +1,7 @@
-#[cfg(all(feature = "compress", feature = "util"))]
+#[cfg(all(feature = "compress", feature = "util", not(target_arch = "wasm32")))]
 pub(crate) mod compress;
 #[cfg(all(feature = "util", not(target_arch = "wasm32")))]
 pub(crate) mod decompress;
 
 #[cfg(target_arch = "wasm32")]
-mod wasm;
+pub(crate) mod wasm;
